requirements.txt
Flask==3.0.2
Werkzeug==3.0.2
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
pytest==8.0.2